adding TargetBombIt, TargetTurttle, TargetLucky, Target8Ammo

+++ Assets/Scripts/Manager's/ScoreManager.cs
@@ -14,7 +14,7 @@ public class ScoreManager : MonoBehaviour
     }
 
 
-    public void AddScore(int value)
+    public void AddScore(float value)
     {
         PlayerData.Instance.AddPlayerScore(value);
         scoreTextUI.text = PlayerData.Instance.PlayerScore.ToString();
